refactor: main 페이지 전용 스크립트를 layout.js로 분리하여 불필요한 API 호출 제거
