+++ index.html
@@ -6,6 +6,7 @@
     <title>L'OCCITANE EN PROVENCE</title>
     <link rel="stylesheet" href="/src/styles/style.css" />
     <script type="module" src="/src/js/main.js"></script>
+    <script type="module" src="/src/pages/main/index.js"></script>
   </head>
   <body>
     <div id="header"></div>
