+++ src/js/main.js
@@ -1,2 +1,2 @@
 import "/src/styles/style.css";
-import "/src/pages/main/index.js";
+import "/src/js/layout.js";
